requirements/test_requirements.txt
-r requirements.txt

pytest==8.2.0
black==24.4.2
flake8==7.0.0
mypy==1.10.0
isort==5.13.2
PyYAML==6.0.1
types-PyYAML==6.0.12